requirements.txt
pandas==2.3.1
requests==2.32.4
python-dotenv==0.21.0

